create subcommand
This adds a cli subcommand to create a database based on the config.
This also reworks the test helper to do some configuration b/c this
command requires a bit of configuration to be in place.

This is only the start and this needs further work at some point.
Some system tests would be nice and it would be good to expand it
to support mysql/mysql2 and sqlite3.

Closes #7.