have an ENV var drive what file the CLI loads to configure Ardb
It defaults to `config/db`.  This gets the Ardb CLI more in-line
with how the Sanford CLI works.  The CLI expects that whatever file
is specified by `ENV['ARDB_CONFIG_FILE']` configures Ardb appropriately.

Moving to this design simplified a lot of things and revealed a
whole bunch of "cruft" in the CLI that just wasn't needed.  Good
times.