adding an `AdapterSpy`; use to test `TestHelpers`
This adds formal tests for the `TestHelpers` methods.  To test the
helper methods in isolation, an `AdapterSpy` class is added
to spy on the adapter calls the test helpers make.

The main goal here is to test the test helpers methods.  The spy
class stuff is more of a side-effect of doing that.