Add a relation spy
This adds a relation spy to Ardb for testing things that
manipulate ActiveRecord relations. This spy will collect
ActiveRecord query expressions and will store them in a
collection for inspection. The spy will also properly fake the
`all` and `count` methods. These will honor limit and offset
values that have been applied. This is tool for testing and can
injected into a system to see what kind of ActiveRecord relation's
it builds.