Expand `Ardb::RelationSpy`, add order values accessors
This updates the `Ardb::RelationSpy` to allow accessing the order
value accessors. These are more `ActiveRecord::Relation` methods.
This also changed the `order` method to add to the `order_values`,
which mimics ActiveRecord's behavior. Finally, this also tweaked
the `limit_value` and `offset_value`. `offset_value` is now
defaulted to `nil` which more correctly mimics ActiveRecord. This
is all to expand the relation spy so it can be used for more
testing scenarios.