Updates to support Ruby 2.0, 2.1 and Rubinius 2.2
This updates Ardb to support Ruby 2.0, 2.1 and Rubinius 2.2. This
didn't require any major changes to Ardb's source. It only
required adding the Rubinius `ruby-sl` to the Gemfile.

This also updates the `Runner` and has it add the working
directory to the `$LOAD_PATH`. This is related to working with
Ruby versions greater than 1.8. Since Ruby 1.9, the working
directory is no longer added to the `$LOAD_PATH`. Thus, the
working directory needs to be added for Ardb commands so the
default isn't useless.