Update `RelationSpy` to support `to_sql`
This updates the `RelationSpy` to support ActiveRecord's `to_sql`
method. This is to support more of ActiveRecord's relation
interface for tools using this method. This method is primarily
intended to support tools using this to create complex SQL queries
with subqueries or possibly unions. This creates a simple string
representing the applied expressions. The string is not actual
SQL.