Move schema dumping to the adapter, setup for sql schema format
This moves the schema dumping logic to the adapter instead of in
the migrate command. This puts it next to the loading schema
logic. It also sets up supporting SQL schema format loading and
dumping, which will have to be done via the adapters. For now,
none of the adapters define sql schema handling and it isn't
used. Eventually, the adapters will check a configuration to see
if they need to use the ruby or sql loading/dumping logic.