Cleanups to the test suite
This updates Ardb's test suite to our latest conventions and
makes sure it is using the latest assert as well. This primarily
renames the `BaseTests` to `UnitTests` but also does a few other
cleanups. This is prep for some new features and changes. This
gets them in the best state possible for this.