Add `schema_format` and use when loading/dumping the schema
This adds a `schema_format` configuration option and uses it in
the schema loading and dumping logic. This allows specifying if
you want to use ruby or SQL for the schema format. This also
establishes the convention for Ardb that it always dumps the ruby
version of the schema and optionally dumps the SQL schema if the
format is set to SQL. It uses the `schema_format` to determine
if it loads the ruby schema file or the SQL schema file.

This changes the nature of the `schema_path` configuration option.
It no longer points to the a single-file, but provides a template
for the schema logic to build a ruby or SQL schema file path. They
do this by appending a ruby or SQL extension onto the value of the
configuration option.