Add PostgreSQL logic for loading/dumping a SQL schema
This adds the logic needed to properly load and dump a SQL schema
using the PostgreSQL adapter. This runs the `pgsql` and `pg_dump`
commands to achieve this. This is tested via stubs to make sure
the commands are what is expected. I've loaded this with an
application using PostgreSQL and made sure the commands work as
expected.