Add scope methods to `RecordSpy`, use `RelationSpy`
This updates the `RecordSpy` to have scope methods and to use
the `RelationSpy` to spy them. This is a common pattern that has
proven to be useful for unit testing scopes when a test DB is not
available. This makes the pattern part of the `RecordSpy` so it
doesn't have to be redone everytime we need to unit test a scope.

This also updates `RelationSpy` to have a `reset!` method. This
is useful because a `RecordSpy`s relation spy will continue to
build up expressions on it. This can be annoying because the
record spy class is typically a global and tests can easily
interfere with one another. This provides a method to reset the
relation spy to avoid these issues.