Add `define_model_callbacks` and `run_callbacks` to `RecordSpy`
This updates `RecordSpy` to support the `define_model_callbacks`
and `run_callbacks` methods. These are used to define custom
callbacks and manually run them.

The `define_model_callbacks` method spies the types of custom
callbacks that are defined. This allows easily testing that a
custom callback is being defined for the record. Furthermore, it
mimics activerecords behavior by adding the class methods for the
callbacks. These are the same as the standard callback methods and
can be tested in the same manner.

The `run_callbacks` method also spies what callbacks are manually
run and yields if a block is passed. This mimics activerecords
behavior and allows easily testing that a callback was manually
run.