migrate cli: show any error backtraces
This is just a minor cleanup.  I ran into a cryptic migration error
and had no context.  I updated this and reran my migration and the
backtrace confirmed it was coming from activerecord.  This helped
tip me off to a typo in my migration.

Now, I'd prefer activerecord throw a less cryptic error, but that
is a separate issue.  This at least helped me figure things out.