Add `Ardb::UseDbDefault` mixin
This adds a use db default mixin which will allow columns to be
defaulted by the db. This is specifically needed to make use of
defaulting a non primary key column using a postgres sequence.

Firstly, activerecord doesn't really use defaults that are
configured on columns. For simple defaults, like a literal value,
activerecord does the actual defaulting when you instantiate an
instance of a record. Otherwise all columns have a `nil` value.
When these are saved, activerecord sends values for all columns
which keeps the DB from defaulting them.

This allows the DB to default values by keeping activerecord from
sending them when it runs a sql `INSERT`. To do this, the mixin
adds an around create callback that will check if a column has
been changed. If it hasn't then we remove it from the the
attributes hash. If this is done, activerecord won't send it in
its sql `INSERT`.

After the record has been created though, it still won't have
values for DB defaulted attributes. Activerecord doesn't reload
all values by default. To fix this, for the attributes that we let
the DB default, we `pluck` their values from the DB. This ensures
we have them after the model is created.

As part of this effort, this also extends the relation spy to have
`pluck` and allow providing values for it to return.