Add `HasSlug` mixin
This adds a `HasSlug` mixin that provides automatic slugging to
active record models. This allows creating readable slugs for
easily.

The mixin adds a `has_slug` class method when included. This can
be used to add the automatic slugging behavior and configure it as
needed. The `has_slug` requires a source proc for configuring
how the slug is generated. It uses this to generate the initial
slug.

Once `has_slug` has been called, it adds validations and callbacks.
It adds a validates presence on update for the configured slug
attribute. Its on update because slugs are generated after the
model has been saved (so the `id` can be used). It also adds a
validates uniqueness of for the slug attribute. This can be
further customized by specifying a unique scope option when
calling `has_slug`. It adds an after create and after update
callback. Both of these methods generate a slug. This is done on
create and update to initially generate the slug and also to make
sure the slug stays a valid slug. If the slug attribute is already
set, the callbacks will attempt to slug the existing string to
make sure it stays a valid slug. Otherwise, the source proc is
used to get an initial string.

This also updates the `UseDbDefault` mixin to prefix its config
method with `ardb`. This should've been done when it was
implemented but was accidentally missed.