Fix has slug when the source is already a valid slug
This fixes a bug with `HasSlug` that caused it to not set its slug
when the source was already a valid slug. This was missed when
initially adding the `HasSlug` mixin.

The tests always assumed slugging the source produced something
different than the source's original value. The problem was that
the source value was being used to see if it matched the generated
slug instead of the slugs attributes current value. In the tests,
the source is setup so it will never match the generated slug
because we are testing other behavior like using the preprocessor
and allowing underscores. This adds a test that makes the source
and the generated slug match but the slug attribute be different
and ensures it sets the slug correctly.

This also fixes a minor bug with the slug tests using non-word
chars. I meant to use the ASCII range between '[' and '`' but
accidentally concatenated them. This caused the tests to fail
because they expected to be using single char seperators.