Allow passing a custom separator to `has_slug`
This updates the has slug mixin to allow passing a custom
separator. This can be used to further customize the slugs it
generates. The default separator is still a dash (`-`).

This also fixes the spelling of "seperator" to the correct
spelling, "separator".

This also adds default constants for the attribute and
preprocessor. This doesn't change any functionality but it makes
it simple to change the defaults.