Move the logic for Migrating a DB into Adapter::Base
This moves the logic for migrating a DB into Adapter::Base so that
it may be called independently of the command line.  Though the migration
doesn't directly use the Adapter, it is a common place for it to live
so that all Adapters can migrate the db in a similar manner as they
create or drop the db.

This also updates the TestHelpers so that apps do not need to directly
interact with the Adapter to drop, create, migrate, or reset the db.
This used the existing pattern of only running a DB command once per
call, but still allowing users to force the call using the bang version
of the method.