Update Adapter based Commands to use AdapterSpy in testing
This updates the testing of Ardb Commands that use the Adapter to
use the Ardb::AdapterSpy.  This allows the testing of `run` behavior.
This also updates those commands to optionally accept an output and
error IO, and use those IO's rather than `$stdout` and `$stderr`.  This
was done so that in the test environment we can avoid writing to the
console when running these Commands.

This also updates the AdapterSpy to include methods for spying the
`dump_schema` command as well as helpers to determine if a command
was run.