Add `DefaultOrderBy` mixin
This adds a default order by mixin that provides the logic for
automatically defaulting a record to the next available "order-by"
value. This is intended to make it simple to easy order a set of
records when they are initially created.

The mixin adds a before validation callback that will run a query
to check for the current max order by value in the DB. It will then
set its order by value to the current max + 1. This makes records
order by default in the order they are created.

The mixin also allows providing a custom scope proc. This will be
used when determining the current max order by value. This can be
used to group or scope order by values. For example, if all the
records have a grouping and each grouping of records should be
ordered separately then the scope can be used to scope the current
max using the records grouping.

The mixin only sets the order by before validation on create when
it hasn't been given a custom value. The order by can be reset to
the next highest value at anytime though by calling the
`reset_order_by` method. This can be used with callbacks to reset
a records order by when an attribute changes and, for example, it
switches from one grouping to another.