Allow not dumping schema when migrating db
This updates the migrate command to not always dump the schema. If
the no schema env var is set then it won't dump the schema after it
migrates. This is convenient when deploying migrations and using
a sql schema. The bins needed to dump the SQL schema may not be
available on the machine running the migration. Also, the schema
should have been dumped when the migration was run locally. It
doesn't need to be dumped again by the server.