rework CLI 'connect' sub-command
This reworks the 'connect' CLI subcommand to put it in the new CLI
framework.  This also enhances the tests to more thoroughly test
the stdout/stderr the command produces.

This is part of porting the CLI sub-commands into the new framework.

Note: this also updates the framework to handle a `CommandExitError`s.
These are used by sub commands to force the CLI to error-exit with
a non-zero status but not output any debug info or help info.