rework cli sub-commands to not have an `init` method
This method is a bit superfluous and can be incorporated into the
run method. The goal is to not have to remember to call both init
and run just the actually run cli commands.

This removes the method and updates the tests appropriately. This
is part of setting up our sub-command framework.