rework CLI 'generate' sub-command
This reworks the 'generate' CLI subcommand to put it in the new CLI
framework. This also enhances the tests to more thoroughly test
the stdout/stderr the command produces.

In the new framework, this is now modeled as a `generate-migration`
sub-command.  A migration is the only "thing" you can generate right
now and this change allows generating migrations to take distinct
options from generating other things without having to do sub-commands
on sub-commands.

Note: this also models a migration so that the save logic isn't
bound up in some CLI command and it can in theory be called in other
contexts.

This is part of porting the CLI sub-commands into the new framework.