rework how CLI initializes Ardb and handles clirb exceptions
This makes the CLI in general init Ardb without establishing a
connection regardless of the command - all commands should expect
Ardb to be configured and initialized.

This also updates the sub-commands to not treat clirb parse errors
as command errors.  This breaks the built in clirb error handling
the CLI does.

These were things that were discovered while using the new CLI in
an application.