Add requiring all of activerecord to init
This adds requiring all of activerecord to ardb's init. This avoids
uninitialized constant errors that happen in threaded environments.
Activerecord uses ruby's autoload which isn't threadsafe. This
is intended to avoid autoload by manually requiring all the files
needed. `Ardb.init` should be run when an app is initialized before
any threads are created.

This also adds a script for determining the files that need to be
manually required. This is a tool to use when activerecord is
updated to ensure we are continue requiring any new files that get
added (or stop requiring old files that are removed).