has slug: allow multiple slug'd attributes per record
The goal here is to have multiple attributes on a given record that
use the has slug logic to drive their values.  Very little is changed
from the original API - you still call `has_slug` as before and the
attribute name is defaulted as before if no custom attr name is given.

The biggest changes are the following:

* you now have the possibility of multiple has-slug configs per record,
  therefore the class method has been pluralized to `ardb_has_slug_configs`
  and is a hash of configs keyed by their attr name
* the `:attribute` key is no longer stored in the has-slug config -
  it is now given everywhere you need to lookup a config or the default
  is assumed
* calling `reset_slug` now assumes the default attr - optionally call
  it with the appropriate attr name to reset the non-default attrs

One implementation note: I didn't change the tests except for the
changes outlined above.  I got everything passing by changing how
the config was referenced in the tests and changing to pluralized
method names where needed.  Once everything passed with those
superficial changes, I then added additional tests to verify calling
`reset_slug` with and without a given attr name to verify this new
behavior case.

Update: As follow up to code review, I updated the tests further to call
`has_slug` multiple times, both with a custom attr like before and
with the default attr.  I then updated the tests to show that updating
the slugs updates all has-slug attrs configured (per feedback from
Collin).  As part of this change, I decided to take out the default
values on the internal `Slug` util class.  It is inconsistent that the
previous defaults used the higher order has-slug default separator but
used a custom default preprocessor.  This inconsistency showed when I
updated the tests and I had to do custom values to get a default exp
slug.  I decided to just remove the magic and make you pass in the
separator and preprocessor for all calls (including the tests).  In
the live code we always pass the separator and preprocessor anyway.