hotfix: have HasSlug always dup the given source value before processing
This ensures that the source value is never modified in place when
performing the slug logic.  We want to allow users to specify, for
example, `:capitalize!` (which modifies its receiver) as a preprocessor
and not modify the source attribute.

This came up using `:to_s` as a preprocessor in an app - b/c `to_s`
returns the string itself (as opposed to the default `:downcase`
which returns a new string), the source attribute was getting modified.