Add `pool` and `checkout_timeout` options to ardb db config
This exposes a `pool` and `checkout_timeout` option for the ardb
db config. This allows changing activerecord's pool size and the
timeout it uses for checking out a new connection from the pool.
These typically aren't set unless the process is using multiple
threads. This allows more control over the activerecord
configuration in a threaded environment. This also makes it
possible to set all of the common activerecord connection options
using ardb.