Add `DbTests` assert context
This adds a `DbTests` assert context which can be used to run
tests in a transaction that is rolled back. This is being added
to help with tests that need to interact with a DB. The DB is a
global state for a test suite and this handles automatically
resetting it. This makes it so tests don't have to handle cleaning
up data they created or deal with a dirty global state. Finally,
this also can improve test speed by never committing anything to
the DB.

The `DbTests` relies on the assert testing library but assert is
not being brought in as a formal dependency. The `DbTests` is an
optional helper that doesn't have to be used. Thus, the user of
ardb is required to bring in ardb if they want to use `DbTests`.
If there are ever more test helpers (for assert or other testing
libraries) then we will probably pull this into an ardb assert
combination gem which will require both dependencies.