Add `escape_like_pattern` helper method to adapters
This adds an `escape_like_pattern` helper method to adapters. This
can be used to escape wildcard characters in `LIKE` patterns so
users don't accidentally use wildcards. A typical use-case is
using a user supplied string to run a SQL query using `LIKE`. The
user may use characters like `%` and `_` which have special
meaning unless they are properly escaped. This helper can be used
to escape the user input so if a wildcard was used it is treated
literally. The helper also escapes the escape char which ensures
it is also treated literally. More than likely, a user intends the
characters to be treated literally and not have their special
meaning.

This is being added as an adapter helper method because its
possible that each adapter may have custom escape logic for like
patterns. The current set of adapters we support all use the same
logic and most likely, all SQL engines will be very similar for
like patterns. In the case we add more helper methods for escaping
adapter specific features then I would want all of our helpers to
be consistently implemented. Thus, I felt the best place to add the
helper method was to the adapter because it allowed the most
flexibility and allowed consistently implementing additional
helpers.

Note: Technically, a user unexpectedly being able to use a wildcard
character is SQL injection. It is most likely not nearly as
dangerous as most SQL injection issues but this also helps keep
users from getting access to data they aren't intended to be able
to access.