Remove `db_file` as a config option
This removes the `db_file` as a config option and adds the ability
to require it relative to the current working directory. This is
part of an effort to remove ns-options from the config and is also
part of updating Ardb to work with ruby 1.9+.

First, this removes the `db_file` as a config option and switches
to always using the env var. This is part of removing ns-options
because this reduces coupling to the `Config` and is one less
config option that has to be ported to a non ns-options
implementation. The `db_file` was confusing as a config option
because practically the `db_file` always needed to be configured
outside of the actual db file to be of any use. The design of
Ardb is that all the common configuration can be contained in the
db file except the db file configuration itself. This avoids the
confusion of having one option that should always be set outside of
the db file while all others are set in the db file.

Next, this also updates how the db file gets required to account
for changes in ruby 1.9+. Ruby 1.9+ doesn't automatically add the
current directory as a load path. This causes the default Ardb db
file `config/db` to not work unless the current directory is
manually added to the load paths. Typically this errors when
running the `ardb` bin because it can't find the db file and the
load paths can't be manually updated. To account for this, the db
file require will now fall back to trying to require the db file
via the current directory after it tries requiring it via the load
paths.

Finally, this updates the tests for `Ardb.init` to provide more
coverage and also adds some notes for future changes.