Update `Config` paths to not be ns-option options
This updates the `Config` root path, migrations path and schema
path to not be ns-options options. This is part of the effort to
remove ns-options.

This removes the root path, migrations path and schema path as
options and adds them as simple attributes to the `Config`. This
required changing the `Config` from being used as a singleton to
instead build an instance of the `Config` for Ardb's main config.
This allows using the initialize of the `Config` class to default
the paths and lets us add them as attributes. This also helps
reduce coupling to Ardb's config by allowing the tests to switch
out Ardb's config with a different instance and not have to worry
about resetting the config singleton.

The migrations path and schema path have also been updated to allow
absolute paths if desired. Previously these always had to be a
relative path to the root path. Now the root path can be ignored
and the migrations and schema path can be set to use absolute
paths. This makes Ardb more flexible to different use-cases. This
also makes the root path no longer required. Previously Ardb would
error when init if the root path wasn't set but with these changes
that is no longer needed. This also allowed removing the `RootPath`
helper class that was used with migrations and schema path options
because it's not being used and `File.expand_path` is providing the
functionality it previously did.

Finally, this includes a cleanup to the sqlite adapter to use
`File.expand_path` for its database file path. This provides the
same functionality but is much simpler.