Remove ns-options from `Config`
This removes ns-options from the `Config` class and simplifies the
`Config`. This is being done for a number of reasons. Ardb config
doesn't need to leverage all of the functionality of ns-options so
its overkill to use it for its config. This also simplifies the
upgrade path to a newer ruby because ns-options is not currently
compatible with it. Also by removing a dependency this decouples
and makes the maintenance of ardb easier.

This changes the `Config` class to no longer include ns-options
and changes its options to simple attribute reader/writers and
no longer be accessed via a `db` namespace. The activerecord
attrs are simple accessors and no special processing or type
casting is done. This allows users to have essentially direct
access to the activerecord options. This lets them configure
activerecord however they normally would.

The logger and schema format options were also updated to be
simple accessors. The logger is now defaulted to a stdout ruby
logger and is no longer required. This allows setting the
activerecord logger to `nil` to disable logging if desired. The
schema format is still defaulted to ruby and attempts to cast any
new value to a symbol.

Next, this renames the `db_settings` method to
`activerecord_connect_hash`. This is to make it clear what this
method is intended for and how it should be used. This helps to
avoid coupling to this method which will allow us to change it
freely to whatever activerecord wants.

This also adds a `validate!` method to the `Config` and removes
it from the `Ardb` module. ns-options previously provided the
validation via its `required_set?` method. The `validate!` method
checks if its adapter and database are set, which is what was
previously done by ns-options. In addition, the schema format is
now validated to make sure it matches a known value which avoids
strange errors using ardb features.