Hotfix: Move `Config` tests into ardb tests file
This moves the `Config` unit tests into the ardb tests file. This
is consistent with the `Config` class implementation which is in
the `Ardb` module file. This keeps things consistent and makes the
tests match the implementation. This doesn't change any of the
tests themselves, just where they are organized.