Adapter cleanups and add dump ruby schema tests
This updates the base adapter to make its readers instance
methods, consistently orders its methods and expands its equality
operator to use its configs. This makes the readers lazy to be
consistent with the other adapters and to avoid unecessary
processing. An adapter doesn't need to determine its schema paths
if it's never used to dump or load a schema. The equality operator
can now be more reliably used to test that two adapters match.

This also includes unit tests for the `dump_ruby_schema` method
which were missing. These most likely weren't tested because we
didn't have good stubbing tools with assert at that time. Now that
we do, this can be tested. This also moves the fake schema that is
used when in the `load_ruby_schema` tests to the `test/support`
folder instead of the `tmp` folder. This matches our latest
conventions and ensures we don't think it actually is a temporary
file and remove it.

Finally, this removes additional gemfiles that aren't being used
from the `tmp` folder. This was setup for testing each adapter
individually but never came to pass. We will most likely handle
testing the specific adapters different when/if we decided to do
that.