Pass `Config` to adapters, decouple from `Ardb.config`
This updates the adapters to expect a `Config` instance to be
passed to them when they are initialized instead of always
using `Ardb.config`. This decouples them from Ardb having to be
init which allows the test suite to not have to init ardb in the
`test/helper`.

This updates the `Adapter` class changing it from a singleton to a
module. This simplifies it and removes the overhead of demetering
method calls to the singleton instance. This also removes it's
reset behavior which was only being used in the `Ardb` module
tests. Ardb itself can't really be "reset" because a db file can
only be required once and a reset interface hasn't been provided
for the `Config`. Furthermore, Ardb can't reliably reset
activerecord after a connection has been established. Because of
all of this I chose to simplify the interface and not worry about
testing the reset behavior. This also updates the `Adapter` to
throw a nice error message in the case the adapter isn't supported
by ardb.

This also changes the CLI and migration helpers to build an adapter
using the new `Adapter` interface. This ensures they throw the same
nice error message that initializing ardb throws.

Finally, this removes the initialization of ardb in the
`test/helper` which allows testing the ardb db file env var default
value. This also adds `Adapter` tests which were missing for the
previous implementation.