Simplify and expand `AdapterSpy`, use in migration helpers tests
This simplifies and expands the `AdapterSpy`. This was noticed
while working on decoupling adapters from the global `Ardb.config`.
This simplifies the `AdapterSpy` to just be an adapter (inherit
from the base adapter) that overwrites the adapter methods to be
spy methods. This also expands it to have fake foreign key sql
methods which allows using it in the migration helpers tests.

The `AdapterSpy` now inherits from `Adapters::Base`, ensuring it
has the same interface as a real adapter. It still overwrites the
adapter methods to spy on them as it did before. This removes it
as a module and dynamic class generator. In hindsight, these didn't
make sense as features. None of the tests used its ability to
add a dynamic interface to an adapter class. Furthermore, it
doesn't make good sense for a tool that works with any adapter to
expect an interface that isn't common to all adapters. I also
checked in code that uses Ardb and didn't find the `AdapterSpy`
being used. This matches up with the commit messages relating to
it which mention it being added for the test helpers tests and then
later being used for the cli tests. This also includes a minor
cleanup to the adapter spy to orders its methods so they match
the order in the base adapter.

This also changes the migration helper tests to use the adapter
spy instead of relying on the `Ardb.config` to be configured for a
specific adapter. Previously, the `Ardb.config.adapter` needed to
be configured to use postgres so that the foreign key sql tests
would pass. The test suite is still configuring the adapter to be
postgres for all unit tests because other tests still depend on it
but it will be removed in a future commit.