Hotfix: Add `min_messages` as activerecord attribute to `Config`
This adds `min_messages` to the activerecord attributes that the
`Config` knows about. This is needed to control postgres output
level. This was previously being done because ns-options allows
writing dynamic options that aren't defined up-front. With
ns-options being gone this needs to be added as an attribute
accessor to the `Config`. Since this is only used by postgres, its
not defualted and isn't required.