Add shim that adds support for postgres json columns to activerecord
This adds a helper file `ardb/pg_json` that can be required to
modify activerecord so that it will support postgres json columns.
This is needed for activerecord 3.2 because it doesn't support
json columns by default. This isn't required by default and should
only be used when using ardb with a postgres adapter.

This also sets up having adapter specific system tests. These were
needed to properly test the shim and make sure it modified
activerecord as expected. This only sets up postgres system tests
but this could be expanded to the other adapters and used to test
their methods. The pg json migrations are separated from regular
migrations because pg json logic is explicitly tested and there
isn't any reason for the pg json or (if there were any) postgres
system tests to have to run each others migrations.

This is a similar pattern to how an application that uses Ardb is
tested but has a few differences. Ardb tests need to test multiple
adapters and its global state. Because of this, Ardb resets the DB
before each db test. In an application, the db can usually be reset
once at the beginning of the all of the tests because the Ardb
configuration and adapter don't change while running the test
suite. Ardb also has to test creating, migrating and dropping
databases which is not someting that would normally be done in an
application's test suite. All of this makes it so the DB needs to
be reset before each test. Since Ardb should have a relatively
small number of system tests the performance overhead of doing this
is not as painful as it would be in a much larger test suite for
an application.