Rework CLI and commands to follow ggem's pattern
This reworks the CLI and commands to follow ggem's pattern for
having a CLI with sub commands. This expands the CLI `--help`
to list all the commands with a summary message explaining them.
This is friendlier and makes it easy to see what commands are
available using the CLI.

The commands now inherit from a `ValidCommand` and are used in
conjunction with a `CommandSet` by the CLI. The `ValidCommand`
commonizes the basics of building and running a command which
avoids having to repeat the logic in each command and keep them
consistent. The `CommandSet` is built by the CLI instead of a hash
and handles mapping a command bash arg with the command class.
This allows the CLI to print out a list of all the commands and
their summary which explains what they do.

In addition to this, the args are now passed to the commands run
method and not when its initialized. This allows makes it so the
commands don't have to rebuilt for each new set of args which
simplifies testing and allows the `CommandSet` to build instances
of the commands when they are added.

This also adopts ggem's organization for its cli, cli.rb and its
commands. This keeps the gems consistent and allows us to more
easily apply best practices to both of them as needed.